+++ vulkano/src/lib.rs
@@ -110,7 +110,7 @@ unsafe impl<T: ?Sized> SafeDeref for Box<T> {
 /// Gives access to the internal identifier of an object.
 pub unsafe trait VulkanObject {
     /// The type of the object.
-    type Object;
+    type Object: vk::Handle;
 
     /// Returns a reference to the object.
     fn internal_object(&self) -> Self::Object;